refactor(era): Replace `EraLocalDirectoryStream` struct with a `read_dir` function
