+++ crates/era-downloader/tests/it/main.rs
@@ -1,7 +1,7 @@
 //! Root module for test modules, so that the tests are built into a single binary.
 
-mod dir;
 mod download;
+mod fs;
 mod list;
 mod stream;
